Move URLRequestConstant to specific file
